% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{add_instance_fleet}
\alias{add_instance_fleet}
\title{Adds an instance fleet to a running cluster}
\usage{
add_instance_fleet(ClusterId, InstanceFleet)
}
\arguments{
\item{ClusterId}{The unique identifier of the cluster.}

\item{InstanceFleet}{Specifies the configuration of the instance fleet.}
}
\description{
Adds an instance fleet to a running cluster.
}
\details{
The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x.
}
\examples{

}
