% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{add_instance_groups}
\alias{add_instance_groups}
\title{Adds one or more instance groups to a running cluster}
\usage{
add_instance_groups(InstanceGroups, JobFlowId)
}
\arguments{
\item{InstanceGroups}{Instance groups to add.}

\item{JobFlowId}{Job flow in which to add the instance groups.}
}
\description{
Adds one or more instance groups to a running cluster.
}
\examples{

}
