% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{cancel_steps}
\alias{cancel_steps}
\title{Cancels a pending step or steps in a running cluster}
\usage{
cancel_steps(
  ClusterId = "string",
  StepIds = list(
    "string"
  )
)
}
\arguments{
\item{ClusterId}{The \code{ClusterID} for which specified steps will be canceled. Use RunJobFlow and ListClusters to get ClusterIDs.}

\item{StepIds}{The list of \code{StepIDs} to cancel. Use ListSteps to get steps and their states for the specified cluster.}
}
\description{
Cancels a pending step or steps in a running cluster. Available only in Amazon EMR versions 4.8.0 and later, excluding version 5.0.0. A maximum of 256 steps are allowed in each CancelSteps request. CancelSteps is idempotent but asynchronous; it does not guarantee a step will be canceled, even if the request is successfully submitted. You can only cancel steps that are in a \code{PENDING} state.
}
\examples{

}
