% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{delete_security_configuration}
\alias{delete_security_configuration}
\title{Deletes a security configuration}
\usage{
delete_security_configuration(Name)
}
\arguments{
\item{Name}{The name of the security configuration.}
}
\description{
Deletes a security configuration.
}
\examples{

}
