% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{describe_cluster}
\alias{describe_cluster}
\title{Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on}
\usage{
describe_cluster(ClusterId)
}
\arguments{
\item{ClusterId}{The identifier of the cluster to describe.}
}
\description{
Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on.
}
\examples{

}
