% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{describe_security_configuration}
\alias{describe_security_configuration}
\title{Provides the details of a security configuration by returning the configuration JSON}
\usage{
describe_security_configuration(Name)
}
\arguments{
\item{Name}{The name of the security configuration.}
}
\description{
Provides the details of a security configuration by returning the configuration JSON.
}
\examples{

}
