% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{list_bootstrap_actions}
\alias{list_bootstrap_actions}
\title{Provides information about the bootstrap actions associated with a cluster}
\usage{
list_bootstrap_actions(ClusterId, Marker = NULL)
}
\arguments{
\item{ClusterId}{The cluster identifier for the bootstrap actions to list.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Provides information about the bootstrap actions associated with a cluster.
}
\examples{

}
