% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{list_clusters}
\alias{list_clusters}
\title{Provides the status of all clusters visible to this AWS account}
\usage{
list_clusters(CreatedAfter = NULL, CreatedBefore = NULL,
  ClusterStates = NULL, Marker = NULL)
}
\arguments{
\item{CreatedAfter}{The creation date and time beginning value filter for listing clusters.}

\item{CreatedBefore}{The creation date and time end value filter for listing clusters.}

\item{ClusterStates}{The cluster state filters to apply when listing clusters.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Provides the status of all clusters visible to this AWS account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.
}
\examples{

}
