% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{list_instance_fleets}
\alias{list_instance_fleets}
\title{Lists all available details about the instance fleets in a cluster}
\usage{
list_instance_fleets(
  ClusterId = "string",
  Marker = "string"
)
}
\arguments{
\item{ClusterId}{[required] The unique identifier of the cluster.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Lists all available details about the instance fleets in a cluster.
}
\details{
The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.
}
\examples{

}
