% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{list_instance_groups}
\alias{list_instance_groups}
\title{Provides all available details about the instance groups in a cluster}
\usage{
list_instance_groups(ClusterId, Marker = NULL)
}
\arguments{
\item{ClusterId}{The identifier of the cluster for which to list the instance groups.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Provides all available details about the instance groups in a cluster.
}
\examples{

}
