% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{list_security_configurations}
\alias{list_security_configurations}
\title{Lists all the security configurations visible to this account, providing their creation dates and times, and their names}
\usage{
list_security_configurations(
  Marker = "string"
)
}
\arguments{
\item{Marker}{The pagination token that indicates the set of results to retrieve.}
}
\description{
Lists all the security configurations visible to this account, providing their creation dates and times, and their names. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListSecurityConfigurations calls.
}
\examples{

}
