% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{modify_instance_fleet}
\alias{modify_instance_fleet}
\title{Modifies the target On-Demand and target Spot capacities for the instance fleet with the specified InstanceFleetID within the cluster specified using ClusterID}
\usage{
modify_instance_fleet(ClusterId, InstanceFleet)
}
\arguments{
\item{ClusterId}{The unique identifier of the cluster.}

\item{InstanceFleet}{The unique identifier of the instance fleet.}
}
\description{
Modifies the target On-Demand and target Spot capacities for the instance fleet with the specified InstanceFleetID within the cluster specified using ClusterID. The call either succeeds or fails atomically.
}
\details{
The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.
}
\examples{

}
