% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_package.R
\docType{package}
\name{paws.emr-package}
\alias{paws.emr}
\alias{paws.emr-package}
\title{paws.emr: Amazon Elastic MapReduce}
\description{
Amazon EMR is a web service that makes it easy to
    process large amounts of data efficiently. Amazon EMR uses Hadoop
    processing combined with several AWS products to do tasks such as web
    indexing, data mining, log file analysis, machine learning, scientific
    simulation, and data warehousing.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
