% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{remove_auto_scaling_policy}
\alias{remove_auto_scaling_policy}
\title{Removes an automatic scaling policy from a specified instance group within an EMR cluster}
\usage{
remove_auto_scaling_policy(ClusterId, InstanceGroupId)
}
\arguments{
\item{ClusterId}{Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.}

\item{InstanceGroupId}{Specifies the ID of the instance group to which the scaling policy is applied.}
}
\description{
Removes an automatic scaling policy from a specified instance group within an EMR cluster.
}
\examples{

}
