% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.emr_operations.R
\name{describe_step}
\alias{describe_step}
\title{Provides more detail about the cluster step}
\usage{
describe_step(ClusterId, StepId)
}
\arguments{
\item{ClusterId}{The identifier of the cluster with steps to describe.}

\item{StepId}{The identifier of the step to describe.}
}
\description{
Provides more detail about the cluster step.
}
\examples{

}
